Move telemetryState file and add breaking comment
