+++ src/library/telemetryState.ts
